[refacto] footer parsing to specific module
